requirements.txt
markitdown[all]
openai
gradio
python-dotenv
